+++ packages/orq-rc/RELEASES.md
@@ -8909,3 +8909,13 @@ Based on:
 - [typescript v4.9.0-rc.29] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.29 - packages/orq-rc
+
+## 2026-05-07 08:09:13
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.30] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.30] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.30 - packages/orq-rc